feat: add CI/CD, shell completions, clippy fixes
- Add GitHub Actions CI: check, test (3 OSes), fmt, clippy
- Add release workflow for cross-platform builds (5 targets)
- Add shell completions via clap_complete (bash, zsh, fish, powershell)
- Fix clippy warnings: rename from_str to parse, derive Default
- Add rustfmt.toml and .clippy.toml configuration

+++ crates/gog-core/src/timeparse.rs
@@ -78,7 +78,7 @@ fn parse_relative_past(lower: &str) -> Option<DateTime<Utc>> {
     let rest = lower.strip_suffix(" ago")?;
     let (n, unit) = split_n_unit(rest)?;
     let now = Utc::now();
-    let dt = apply_offset(now, -(n as i64), unit)?;
+    let dt = apply_offset(now, -n, unit)?;
     Some(dt)
 }
 
@@ -88,7 +88,7 @@ fn parse_relative_future(lower: &str) -> Option<DateTime<Utc>> {
     let rest = lower.strip_prefix("in ")?;
     let (n, unit) = split_n_unit(rest)?;
     let now = Utc::now();
-    let dt = apply_offset(now, n as i64, unit)?;
+    let dt = apply_offset(now, n, unit)?;
     Some(dt)
 }
